Added analysis script.
First version of an example script that makes plots to locate the percolation transition.

I noticed an error: currently the C module accepts only python int types, not
np.int64, for example. We should use the NumPy C API to correct this and allow
any type of integer to be passed, and to return a np array.