First version of Hamiltonian creation.
Added a function H_hypercube() to hypergraphs.c.

This function creates the Hamiltonian for the Hypercube (with percolation) directly as a NumPy array.

To do this, support for a standalone executable has been removed.

Code needs to be unit tested before production.

Need to add the same support to PXP graph - reuse as much code as possible by splitting things off into helper functions.

Still need to unit test the cluster growth for PXP.