Added hypercube_dijkstra_LC
This function works in a similar way to hypercube_H_LC. We run dijkstra's
algorithm from site 0, then again from the first unvisited node, and so on,
until we have visited the entire HS (or encountered an exit condition).

This works because we can use the same arrays visited[] and distances[] for
the whole HS (as the sub-graphs are not connected), and just copy over the
right part of distances[] at the end, using a systems of labels to mark each
node as belonging to a distinct cluster.

Added unit tests, which have passed.